Explicitly name master branch in Travis icon url
For some reason, when not giving a branch, Travis CI currently seems to return
the icon for the build status of the `tvOS_Xcode_7.1` branch.
This shows a "failing" icon on the masters Readme.

I think this is a bug in Travis CI, that can be worked around by adding
a branch=master parameter to the svg url