Clean up CocoaLumberjack.swift a little
Rename `SwiftLogMacro` to `_DDLogMessage` instead, because
`SwiftLogMacro` is a weird name for this. Leave the old name behind but
marked as deprecated for now; it should be removed later. Give
`_DDLogMessage` a more sane argument ordering as well (and no default
arguments; it's not meant for end users to call).

Stop using string interpolation for the `StaticString`s. The comments
say it avoids an integer overflow warning, but I see no such warning
(and never have).

Rewrite `CurrentFileName()` in pure Swift instead of unnecessarily
bridging to Cocoa.